refactor(ses): extract the identity domain into SesIdentityService
